+++ pkg/store/node_compat_external_test.go
@@ -5,7 +5,7 @@ import "github.com/sparkwing-dev/sparkwing/pkg/store"
 var _ = store.Node{
 	"", "", "", "", nil, "", nil, nil, nil,
 	nil, "", false, nil, nil, 0, 0, 0, "", "", "", "", 0, "",
-	"", "", "", "", "", "", nil, "", 0, "", nil, "", "", "", nil,
+	"", "", "", "", "", "", "", "", nil, "", 0, "", nil, "", "", "", nil,
 	"", nil, 0, 0,
 	nil, nil, 0, 0, 0, "", nil, "", nil, nil, "", 0, 0, 0, "",
 }